// src/services/models/formattedTopHolderResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Swipe Pad API
 * OpenAPI spec version: 0.0.12
 */
import type { UserResponse } from "./userResponse";

export interface FormattedTopHolderResponse {
  readonly walletAddress: string;
  readonly amount: string;
  readonly holdingPercent: string;
  readonly userInfo: UserResponse;
}